requirements.txt
Flask
Flask-Cors
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
mysql-connector-python
python-dotenv
jinja2
